make util easier to read
