+++ src/util/serialization.rs
@@ -1,0 +1,22 @@
+use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
+
+// For serialization, ArrayD uses serde while G1Affine uses ark_serialize.
+// In order to bridge between the two, the following code snippet is used:
+// https://github.com/arkworks-rs/algebra/issues/178#issuecomment-1413219278
+pub fn ark_se<S, A: CanonicalSerialize>(a: &A, s: S) -> Result<S::Ok, S::Error>
+where
+  S: serde::Serializer,
+{
+  let mut bytes = vec![];
+  a.serialize_compressed(&mut bytes).map_err(serde::ser::Error::custom)?;
+  s.serialize_bytes(&bytes)
+}
+
+pub fn ark_de<'de, D, A: CanonicalDeserialize>(data: D) -> Result<A, D::Error>
+where
+  D: serde::de::Deserializer<'de>,
+{
+  let s: Vec<u8> = serde::de::Deserialize::deserialize(data)?;
+  let a = A::deserialize_compressed_unchecked(s.as_slice());
+  a.map_err(serde::de::Error::custom)
+}
